+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * feat(EditorSuggestionMenu): expose suggestion matching options
 * feat(Select): support `item-aligned` position mode
 * feat(components): resolve `defaultVariants` in template logic
+* feat(CommandPalette): add `group-label` slot
 * 
 * fix(Error): support `status` and `statusText` properties
 * fix(ContentSurround): handle RTL mode